style(header): add mobile style to header contacts
