+++ src/components/Icons/EmailIcon.jsx
@@ -1,14 +1,14 @@
 const EmailIcon = () => {
   return (
     <svg
-      width="20"
+      width="24"
       height="20"
-      viewBox="0 0 20 20"
+      viewBox="0 0 24 20"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        d="M16.6667 3.33331H3.33332C2.41666 3.33331 1.67499 4.08331 1.67499 4.99998L1.66666 15C1.66666 15.9166 2.41666 16.6666 3.33332 16.6666H16.6667C17.5833 16.6666 18.3333 15.9166 18.3333 15V4.99998C18.3333 4.08331 17.5833 3.33331 16.6667 3.33331ZM16.3333 6.87498L10.4417 10.5583C10.175 10.725 9.82499 10.725 9.55832 10.5583L3.66666 6.87498C3.5831 6.82807 3.50992 6.7647 3.45156 6.68869C3.39321 6.61268 3.35088 6.52563 3.32714 6.43279C3.3034 6.33995 3.29874 6.24326 3.31345 6.14857C3.32816 6.05388 3.36194 5.96316 3.41272 5.8819C3.46351 5.80064 3.53026 5.73053 3.60893 5.67581C3.6876 5.62109 3.77655 5.58291 3.8704 5.56356C3.96426 5.54422 4.06106 5.54412 4.15495 5.56327C4.24885 5.58242 4.33788 5.62042 4.41666 5.67498L9.99999 9.16665L15.5833 5.67498C15.6621 5.62042 15.7511 5.58242 15.845 5.56327C15.9389 5.54412 16.0357 5.54422 16.1296 5.56356C16.2234 5.58291 16.3124 5.62109 16.3911 5.67581C16.4697 5.73053 16.5365 5.80064 16.5873 5.8819C16.638 5.96316 16.6718 6.05388 16.6865 6.14857C16.7012 6.24326 16.6966 6.33995 16.6728 6.43279C16.6491 6.52563 16.6068 6.61268 16.5484 6.68869C16.4901 6.7647 16.4169 6.82807 16.3333 6.87498Z"
+        d="M21.6 0.399902H2.4C1.08 0.399902 0.012 1.4799 0.012 2.7999L0 17.1999C0 18.5199 1.08 19.5999 2.4 19.5999H21.6C22.92 19.5999 24 18.5199 24 17.1999V2.7999C24 1.4799 22.92 0.399902 21.6 0.399902ZM21.12 5.4999L12.636 10.8039C12.252 11.0439 11.748 11.0439 11.364 10.8039L2.88 5.4999C2.75967 5.43236 2.6543 5.3411 2.57027 5.23165C2.48623 5.1222 2.42527 4.99684 2.39109 4.86315C2.35691 4.72946 2.3502 4.59022 2.37139 4.45387C2.39257 4.31752 2.4412 4.18688 2.51434 4.06986C2.58747 3.95285 2.68359 3.85189 2.79687 3.7731C2.91015 3.6943 3.03825 3.63932 3.17339 3.61146C3.30854 3.58361 3.44794 3.58346 3.58315 3.61104C3.71835 3.63862 3.84656 3.69334 3.96 3.7719L12 8.7999L20.04 3.7719C20.1534 3.69334 20.2817 3.63862 20.4169 3.61104C20.5521 3.58346 20.6915 3.58361 20.8266 3.61146C20.9618 3.63932 21.0898 3.6943 21.2031 3.7731C21.3164 3.85189 21.4125 3.95285 21.4857 4.06986C21.5588 4.18688 21.6074 4.31752 21.6286 4.45387C21.6498 4.59022 21.6431 4.72946 21.6089 4.86315C21.5747 4.99684 21.5138 5.1222 21.4297 5.23165C21.3457 5.3411 21.2403 5.43236 21.12 5.4999Z"
         fill="#D66600"
       />
     </svg>
