+++ src/components/Icons/FacebookIcon.jsx
@@ -1,23 +1,16 @@
 const FacebookIcon = () => {
   return (
     <svg
-      width="25"
-      height="25"
-      viewBox="0 0 25 25"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <rect
-        x="4.57547"
-        y="4.57544"
-        width="15.8491"
-        height="15.8491"
-        rx="7.92453"
-        fill="#F3F3F3"
-      />
+      <rect width="24" height="24" rx="9" fill="#FFFCFC" />
       <path
-        d="M14.3756 13.0037L14.5995 11.5449H13.1998V10.5983C13.1998 10.1992 13.3953 9.81024 14.0222 9.81024H14.6585V8.56828C14.6585 8.56828 14.0811 8.46973 13.529 8.46973C12.3763 8.46973 11.6228 9.1684 11.6228 10.4332V11.5449H10.3415V13.0037H11.6228V16.5301H13.1998V13.0037H14.3756Z"
-        fill="#120000"
+        d="M14.8403 12.7627L15.1793 10.5537H13.0597V9.12028C13.0597 8.51596 13.3558 7.92689 14.3051 7.92689H15.2687V6.04621C15.2687 6.04621 14.3942 5.89697 13.5582 5.89697C11.8127 5.89697 10.6717 6.95496 10.6717 8.87021V10.5537H8.73145V12.7627H10.6717V18.1027H13.0597V12.7627H14.8403Z"
+        fill="#141B2A"
       />
     </svg>
   );
