+++ src/components/Icons/LocationIcon.jsx
@@ -2,13 +2,13 @@ const LocationIcon = () => {
   return (
     <svg
       width="20"
-      height="20"
-      viewBox="0 0 20 20"
+      height="24"
+      viewBox="0 0 20 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        d="M9.66667 8.83333C10.125 8.83333 10.5175 8.67 10.8442 8.34333C11.1708 8.01667 11.3339 7.62444 11.3333 7.16667C11.3333 6.70833 11.17 6.31583 10.8433 5.98917C10.5167 5.6625 10.1244 5.49944 9.66667 5.5C9.20833 5.5 8.81583 5.66333 8.48917 5.99C8.1625 6.31667 7.99944 6.70889 8 7.16667C8 7.625 8.16333 8.0175 8.49 8.34417C8.81667 8.67083 9.20889 8.83389 9.66667 8.83333ZM9.66667 17.1667C7.43056 15.2639 5.76056 13.4967 4.65667 11.865C3.55278 10.2333 3.00056 8.72278 3 7.33333C3 5.25 3.67028 3.59028 5.01083 2.35417C6.35139 1.11806 7.90333 0.5 9.66667 0.5C11.4306 0.5 12.9828 1.11806 14.3233 2.35417C15.6639 3.59028 16.3339 5.25 16.3333 7.33333C16.3333 8.72222 15.7811 10.2328 14.6767 11.865C13.5722 13.4972 11.9022 15.2644 9.66667 17.1667Z"
+        d="M10 12C10.6875 12 11.2762 11.7648 11.7663 11.2944C12.2563 10.824 12.5008 10.2592 12.5 9.6C12.5 8.94 12.255 8.3748 11.765 7.9044C11.275 7.434 10.6867 7.1992 10 7.2C9.3125 7.2 8.72375 7.4352 8.23375 7.9056C7.74375 8.376 7.49917 8.9408 7.5 9.6C7.5 10.26 7.745 10.8252 8.235 11.2956C8.725 11.766 9.31333 12.0008 10 12ZM10 24C6.64583 21.26 4.14083 18.7152 2.485 16.3656C0.829167 14.016 0.000833333 11.8408 0 9.84C0 6.84 1.00542 4.45 3.01625 2.67C5.02708 0.89 7.355 0 10 0C12.6458 0 14.9742 0.89 16.985 2.67C18.9958 4.45 20.0008 6.84 20 9.84C20 11.84 19.1717 14.0152 17.515 16.3656C15.8583 18.716 13.3533 21.2608 10 24Z"
         fill="#D66600"
       />
     </svg>
