+++ src/components/Icons/PhoneIcon.jsx
@@ -9,7 +9,6 @@ const PhoneIcon = () => {
     >
       <path
         fillRule="evenodd"
-       
         d="M2.35638 0.638567C2.5751 0.420178 2.83773 0.250765 3.12688 0.141555C3.41602 0.0323445 3.72507 -0.0141693 4.03356 0.00509601C4.34204 0.0243613 4.6429 0.108966 4.91621 0.253302C5.18952 0.397639 5.42903 0.598413 5.61888 0.842317L7.86263 3.72482C8.27388 4.25357 8.41888 4.94232 8.25638 5.59232L7.57263 8.32982C7.53728 8.4716 7.53919 8.62012 7.57818 8.76095C7.61716 8.90179 7.6919 9.03014 7.79513 9.13357L10.8664 12.2048C10.9699 12.3083 11.0985 12.3831 11.2396 12.4221C11.3806 12.4611 11.5294 12.4629 11.6714 12.4273L14.4076 11.7436C14.7284 11.6634 15.0632 11.6571 15.3867 11.7254C15.7103 11.7936 16.014 11.9344 16.2751 12.1373L19.1576 14.3798C20.1939 15.1861 20.2889 16.7173 19.3614 17.6436L18.0689 18.9361C17.1439 19.8611 15.7614 20.2673 14.4726 19.8136C11.1741 18.653 8.17922 16.7646 5.71013 14.2886C3.23424 11.8198 1.34589 8.82539 0.185131 5.52732C-0.267369 4.23982 0.138881 2.85607 1.06388 1.93107L2.35638 0.638567Z"
         fill="#D66600"
       />
