+++ src/components/Icons/YoutubeIcon.jsx
@@ -1,11 +1,18 @@
-
-
 const YoutubeIcon = () => {
   return (
-    <svg width="25" height="25" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M12.744 4.5C13.278 4.503 14.614 4.516 16.034 4.573L16.538 4.595C17.967 4.662 19.395 4.778 20.104 4.975C21.049 5.241 21.791 6.015 22.042 6.997C22.442 8.557 22.492 11.599 22.498 12.336L22.499 12.488V12.662C22.492 13.399 22.442 16.442 22.042 18.001C21.788 18.986 21.045 19.761 20.104 20.023C19.395 20.22 17.967 20.336 16.538 20.403L16.034 20.426C14.614 20.482 13.278 20.496 12.744 20.498L12.509 20.499H12.254C11.124 20.492 6.398 20.441 4.894 20.023C3.95 19.757 3.207 18.983 2.956 18.001C2.556 16.441 2.506 13.399 2.5 12.662V12.336C2.506 11.599 2.556 8.556 2.956 6.997C3.21 6.012 3.953 5.237 4.895 4.976C6.398 4.557 11.125 4.506 12.255 4.5H12.744ZM10.499 9V16L16.499 12.5L10.499 9Z" fill="#F3F3F3"/>
+    <svg
+      width="30"
+      height="24"
+      viewBox="0 0 30 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M15.3665 0C16.1675 0.00450028 18.1716 0.0240014 20.3018 0.109507L21.0578 0.142509C23.2015 0.243015 25.3436 0.417026 26.4072 0.712544C27.8247 1.11157 28.9378 2.27264 29.3143 3.74573C29.9144 6.08588 29.9894 10.6492 29.9984 11.7547L29.9999 11.9827V12.2438C29.9894 13.3493 29.9144 17.9141 29.3143 20.2528C28.9333 21.7304 27.8187 22.8929 26.4072 23.286C25.3436 23.5815 23.2015 23.7555 21.0578 23.856L20.3018 23.8905C18.1716 23.9745 16.1675 23.9955 15.3665 23.9985L15.0139 24H14.6314C12.9363 23.9895 5.84688 23.913 3.59074 23.286C2.17465 22.8869 1.06008 21.7259 0.683555 20.2528C0.0835171 17.9126 0.00851228 13.3493 -0.000488281 12.2438V11.7547C0.00851228 10.6492 0.0835171 6.08438 0.683555 3.74573C1.06458 2.26814 2.17915 1.10557 3.59224 0.714044C5.84688 0.0855051 12.9378 0.00900056 14.6329 0H15.3665ZM11.9988 6.75042V17.2511L20.9993 12.0008L11.9988 6.75042Z"
+        fill="#FFFCFC"
+      />
     </svg>
-  )
-}
+  );
+};
 
-export default YoutubeIcon 
+export default YoutubeIcon;
