+++ src/components/main/Header/HeaderContacts.jsx
@@ -9,21 +9,9 @@ import YoutubeIcon from '@/components/Icons/YoutubeIcon';
 
 import styles from './Header.module.scss';
 
-const HeaderContacts = () => {
+const HeaderContacts = ({ windowWidth }) => {
   const [show, setShow] = useState(true);
   const [lastScrollY, setLastScrollY] = useState(0);
-  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-  useEffect(() => {
-    function handleResize() {
-      setWindowWidth(window.innerWidth);
-    }
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
 
   useEffect(() => {
     const controlNavbar = () => {
@@ -49,29 +37,31 @@ const HeaderContacts = () => {
       <ul className={styles.contactsList}>
         <li className={styles.contactsListItem}>
           <a
+            className={styles.contactsListLink}
             href="https://www.google.com.ua/maps/place/Kiev+Art+School/@50.4685751,30.3658169,12z/data=!4m10!1m2!2m1!1z0YXRg9C00L7QttC90Y8g0YjQutC-0LTQsCDQstC10YDQvdC40LrRltCy0YHRjNC60L7Qs9C-!3m6!1s0x40d4ce5df69f9943:0x18b52a2948c8f340!8m2!3d50.4543277!4d30.5043019!15sCjTRhdGD0LTQvtC20L3RjyDRiNC60L7Qu9CwINCy0LXRgNC40LrRltCy0YHRjNC60L7Qs9C-kgEKYXJ0X3NjaG9vbOABAA!16s%2Fg%2F121jyvxp?entry=ttuпше "
             target="_blank"
             rel="noopener noreferrer nofollow"
           >
             <LocationIcon />{' '}
-            {windowWidth > 768 ? 'вул. Бульварно-Кудрявська, 2.' : ''}
+            {windowWidth > 1280 ? 'вул. Бульварно-Кудрявська, 2.' : ''}
           </a>
         </li>
         <li className={styles.contactsListItem}>
-          <div className={styles.contactsListItem_PhoneWrapper}>
-            <a href="tel:+380972907940">
-              <PhoneIcon />
-              {windowWidth > 768 ? '+38(097)290-79-40' : ''}
-            </a>
-
-            <a href="tel:+380934560838">
-              {windowWidth > 768 ? ' +38(093)456 08 38' : ''}
-            </a>
-          </div>
+          <a className={styles.contactsListLink} href="tel:+380972907940">
+            <PhoneIcon />
+            {windowWidth > 1280 ? (
+              <div className={styles.contactsListItem_PhoneWrapper}>
+                <span> +38(097)290-79-40</span>
+                <span> +38(093)456 08 38</span>
+              </div>
+            ) : (
+              ''
+            )}
+          </a>
         </li>
-        <li className={styles.contactsListItem}>
-          <a href="mailto:Shkola_2@ukr.net">
-            <EmailIcon /> {windowWidth > 768 ? 'Shkola_2@ukr.net' : ''}
+        <li className={styles.contactsListItem} >
+          <a className={styles.contactsListLink} href="mailto:Shkola_2@ukr.net">
+            <EmailIcon /> {windowWidth > 1280 ? 'Shkola_2@ukr.net' : ''}
           </a>
         </li>
       </ul>
@@ -85,7 +75,7 @@ const HeaderContacts = () => {
           </li>
           <li>
             <a className={styles.socialLink} href="#">
-              <YoutubeIcon width="24" height="24" />
+              <YoutubeIcon />
             </a>
           </li>
         </ul>
